add expense add group and send transaction service added tested with no problem

+++ app/src/main/java/com/example/budgetplus/model/response/ExpenseGroup.kt
@@ -7,5 +7,5 @@ data class ExpenseGroup(
     @SerializedName("groupedExpenses")
     var groupedExpenses: MutableList<GroupedExpense>,
     @SerializedName("total")
-    var total: Double
+    var total: Int
 )
